requirements.txt
Django==1.6
South==0.8.3
django-autoslug==1.7.1
